grading-rubric/pyproject.toml
[project]
name = "grading-rubric-tools"
version = "0.1.0"
description = "Validation helpers for the grading-rubric Cursor skill (RUBRIC.md parsing and structural checks)."
readme = "SKILL.md"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]
addopts = "-q"
